Use token in device unique ID
The token is (assuming it's a UUID) unique to a given Make API instance
across all hubs. It is set when a Maker API instance is created and
never changes.

This should take care of the occasionally entity duplication that
happened when using MACs in the entity IDs.

fixes #53
references #51
references #45