Don't use empty string as event_url
v0.5.x allowed an empty string to be saved as the server_url config
option, and used a use_server_url flag to decide whether to send that
url to the hub. v0.6.x allowed sever_url to be null and dispensed with
the flag. This causes problems with options saved using v0.5.x, as these
man contain an empty string for server_url.

Treat "" as None for server_url.

references #102